requirements.txt
pyaes==1.6.1
PyYAML==6.0.1
requests==2.31.0